add patch as method for days
I wanted to add a method to reorder the food which would require yet another method. To break this pattern I will create a patch method which will take the list as given through and overwrite the last meal list.
this gives the client also to flexibility to do multi adds, deletes, reorders etc without sending every single request to the server